# src/main/java/dev/jpitarch/ctrlgym/verifactu/verifacti-openapi.yml
openapi: 3.0.3
servers:
  - url: https://api.verifacti.com/
info:
  version: 1.0.0
  title: Documentación de la API Verifactu
  contact:
    email: info@verifacti.com
    name: Contacto
    url: https://www.verifacti.com
  description: >

    &nbsp;


    Aquí encontrarás la documentación técnica de nuestra API para enviar
    facturas a la Agencia Estatal de Administración Tributaria (AEAT) de acuerdo

    al sistema Verifactu.


    &nbsp;


    También disponemos de una API para cumplir con los requisitos de TicketBai,
    el análogo a Verifactu en el País Vasco. Se trata de una API

    separada pero con una estructura completamente análoga. La documentación se
    puede encontrar en <a href="https://www.verifacti.com/tb-docs">aquí</a>.


    &nbsp;


    Asimismo, hemos creado una tercera API para poder gestionar los diferentes
    NIFs de manera programática. Para usuarios que solo deben gestionar

    unos pocos NIFs, esto se puede hacer de manera cómoda a través de nuestra
    interfaz. Sin embargo, en ocasiones esto no es práctico y hace falta tener 

    una manera programática de dar de alta y de baja diferentes NIFs. Para ello,
    se puede utilizar nuestra API de gestión de NIFs, cuya documentación se

    puede consultar <a href="https://www.verifacti.com/nifs-docs">aquí</a>. Esta
    API está incluida en el precio y se puede usar por cualquier usuario

    con una suscripción activa.


    &nbsp;


    <div style="display: flex; justify-content: start;">


...


...


...
paths:
...
  /verifactu/status:
    post:
      summary: Estado factura
      description: >
        Este endpoint permite consultar el estado de una factura en el sistema
        de la AEAT.


        &nbsp;


        Las facturas no registradas en la AEAT no aparecerán aquí.

        Para comprobar el estado de registros de facturación no procesados aún o
        rechazados por la AEAT, se debe consultar el endpoint GET `/status`.
      x-codeSamples:
        - lang: cURL
          label: cURL
          source: |-
            curl -X POST 'https://api.verifacti.com/verifactu/status' \
              -H 'Authorization: Bearer <API_KEY>' \
              -H 'Content-Type: application/json' \
              -d '{
                   "serie": "A",
                   "numero": "234634",
                   "fecha_expedicion": "25-05-2026"
                 }'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - serie
                - numero
                - fecha_expedicion
              properties:
                serie:
                  type: string
                  example: A
                  description: Serie de la factura.
                numero:
                  type: string
                  example: '234634'
                  description: Número de la factura.
                fecha_expedicion:
                  type: string
                  pattern: \d{2,2}-\d{2,2}-\d{4,4}
                  example: 25-05-2026
                  description: >-
                    Fecha de emisión de la factura. No puede ser una fecha
                    posterior a la actual.
                fecha_operacion:
                  type: string
                  pattern: \d{2,2}-\d{2,2}-\d{4,4}
                  description: >
                    Fecha de operación de la factura. Es obligatorio incluir la
                    fecha de la operación si ésta se incluyó en el registro de 

                    facturación original y es diferente de la fecha de
                    expedición.
      responses:
        '200':
          description: Estado factura
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/estadoFactura'
        '500':
          description: Error de servidor
    get:
      summary: Estado registro
      description: >
        Este endpoint permite conocer el estado actual de un registro de
        facturación efectuado anteriormente. Inmediatamente después de enviar un
        registro, éste se

        encuentra siempre en estado `Pendiente`, se encola y se procesará
        típicamente en menos de un minuto. Esto es debido a que, en general, la
        AEAT no permite enviar registros

        de facturación en tiempo real si no que obliga a esperar un tiempo entre
        los diferentes envíos.


        &nbsp;


        Es importante entender que, aunque no es lo normal, pueden existir
        múltiples registros de facturación para una misma factura. Por ejemplo,
        una factura creada

        y anulada posteriormente tendría dos registros de facturación diferentes
        (uno para la creación y otro para la anulación). Este endpoint devuelve
        el estado del

        registro de facturación, que no corresponde necesariamente con el estado
        de la factura.


        &nbsp;


        En el entorno de test, los datos relativos a los registros de
        facturación se guardarán un máximo de 90 días.

        Sin embargo, los endpoints de consulta de facturas tendrán los datos
        históricos ya que consultan a la AEAT.
      x-codeSamples:
        - lang: cURL
          label: cURL
          source: >-
            curl -X GET
            'https://api.verifacti.com/verifactu/status?uuid=b018ced3-b362-4494-8776-9eefff1c160c'
            \
              -H 'Authorization: Bearer <API_KEY>'
      parameters:
        - in: query
          name: uuid
          required: true
          schema:
            type: string
          example: b018ced3-b362-4494-8776-9eefff1c160c
          description: >-
            Identificador único del registro recibido como respuesta en el
            momento del envío.
      responses:
        '200':
          description: Estado registro
          content:
            application/json:
              schema:
                type: object
                properties:
                  nif:
                    description: NIF del emisor de la factura.
                    type: string
                  serie:
                    description: Serie de la factura.
                    type: string
                  numero:
                    description: Número de la factura.
                    type: string
                  fecha_expedicion:
                    description: Fecha de emisión de la factura.
                    type: string
                  operacion:
                    description: >
                      Hay tres tipos de operaciones de registro habituales:

                      <ol style="list-style: disc;">
                        <li>Alta: envío normal de una factura. Endpoint: <code>/create</code>.</li>
                        <li>Subsanacion: modificación normal de una factura. Endpoint: <code>/modify</code>.</li>
                        <li>Anulacion: anulación normal de una factura. Endpoint: <code>/cancel</code>.</li>
                      </ol>


                      Existen otras 5 operaciones especiales:

                      <ol style="list-style: disc;">
                        <li>Alta (rechazo previo): modificación tras rechazo previo. Endpoint: <code>/modify</code> con <code>rechazo_previo=X</code>.</li>
                        <li>Alta (rechazo previo de subsanacion): modificación tras rechazo de subsanación. Endpoint: <code>/modify</code> con <code>rechazo_previo=S</code>.</li>
                        <li>Anulación (rechazo previo): anulación tras rechazo previo. Endpoint: <code>/cancel</code> con <code>rechazo_previo=S</code> y <code>sin_registro_previo=N</code>.</li>
                        <li>Anulación (sin registro): anulación sin registro en la AEAT. Endpoint: <code>/cancel</code> con <code>rechazo_previo=N</code> y <code>sin_registro_previo=S</code>.</li>
                        <li>Anulación (rechazo previo sin registro): envío tras rechazo sin registro. Endpoint: <code>/cancel</code> con <code>rechazo_previo=S</code> y <code>sin_registro_previo=S</code>.</li>
                      </ol>


                      Para más información, consultar la <a
                      href='https://www.agenciatributaria.es/AEAT.desarrolladores/Desarrolladores/_menu_/Documentacion/Sistemas_Informaticos_de_Facturacion_y_Sistemas_VERI_FACTU/Sistemas_Informaticos_de_Facturacion_y_Sistemas_VERI_FACTU.html'
                      target='_blank'>documentación AEAT</a>.
                    type: string
                  estado:
                    description: |
                      Estado del registro de facturación:
                      <ol style="list-style: disc;">
                        <li>Pendiente: Registro encolado y no procesado aún.</li>
                        <li>Correcto: Registro procesado correctamente por la AEAT.</li>
                        <li>Aceptado con errores: Registro aceptado con errores por la AEAT. Se requiere enviar un registro de subsanación o emitir una rectificativa.</li>
                        <li>Incorrecto: Registro considerado incorrecto por la AEAT. Se requiere enviar un registro de subsanación con <code>rechazo_previo=S</code> o <code>rechazo_previo=X</code> o emitir una rectificativa.</li>
                        <li>Duplicado: Registro no aceptado por la AEAT por existir un registro con el mismo (serie, numero, fecha_expedicion).</li>
                        <li>Anulado: Registro de anulación procesado correctamente por la AEAT.</li>
                        <li>Factura inexistente: Registro de anulación no aceptado por la AEAT por no existir la factura.</li>
                        <li>No registrado: Registro rechazado por la AEAT.</li>
                        <li>Error servidor AEAT: Error en el servidor de la AEAT. Se intentará reenviar el registro de facturación de nuevo.</li>
                      </ol>
                    type: string
                  url:
                    description: URL de verificación del código QR.
                    type: string
                  qr:
                    description: Código QR en base 64 que contiene la url.
                    type: string
                  codigo_error:
                    description: >-
                      Código de error tal y como aparece en el sistema de la
                      AEAT.
                    type: string
                  mensaje_error:
                    description: >-
                      Descripcion del error tal y como aparece en el sistema de
                      la AEAT.
                    type: string
                  estado_registro_duplicado:
                    description: >-
                      El estado del registro en caso de que el estado del
                      registro sea `Duplicado`.
                    type: string
                example:
                  nif: A15022510
                  serie: A
                  numero: '34547'
                  fecha_expedicion: 25-05-2026
                  operacion: Alta
                  url: >-
                    https://prewww2.aeat.es/wlpl/TIKE-CONT/ValidarQR?nif=A15022510&numserie=A234634&fecha=25-05-2026&importe=242
                  qr: jBBWRw0KGgoABAANSUhEAH0CAIAAABE...
                  estado: Correcto
        '404':
          description: Registro no encontrado
        '500':
          description: Error de servidor
...
components:
  schemas:
    estadoFactura:
      type: object
      properties:
        uuid:
          description: Identificador único del último registro aceptado para esta factura.
          type: string
          example: b018ced3-b362-4494-8776-9eefff1c160c
        nif_emisor:
          description: NIF del emisor de la factura.
          type: string
          example: A15022510
        num_serie:
          description: Serie y número de la factura.
          type: string
          example: A34547
        fecha_expedicion:
          description: Fecha de emisión de la factura.
          type: string
          example: 25-05-2026
        subsanacion:
          description: Indica si la factura se ha subsanado (S) o no (N).
          enum:
            - 'N'
            - S
          type: string
          example: 'N'
        rechazo_previo:
          type: string
          enum:
            - 'N'
            - S
            - X
          example: 'N'
          description: >
            Este campo sirve para indicar si la factura que se quiere subsanar
            fue rechazada previamente por la AEAT.
        tipo_factura:
          type: string
          example: F1
          description: |
            Los diferentes tipos de factura son:
            <ol style="list-style: disc;">
              <li>F1: Factura (Art. 6, 7.2 Y 7.3 del RD 1619/2012)</li>
              <li>F2: Factura simplificada y facturas sin identificación del destinatario (Art. 6.1.D RD 1619/2012)</li>
              <li>R1: Factura rectificativa (Art 80.1 y 80.2 y error fundado en derecho)</li>
              <li>R2: Factura rectificativa (Art. 80.3)</li>
              <li>R3: Factura rectificativa (Art. 80.4)</li>
              <li>R4: Factura rectificativa (Resto)</li>
              <li>R5: Factura rectificativa en facturas simplificadas (Art. 80.3)</li>
              <li>F3: Factura emitida en sustitución de facturas simplificadas facturadas y declaradas</li>
            </ol>
        descripcion:
          type: string
          description: Descripción de la operación.
          example: Descripcion de la operacion
        lineas:
          type: array
          description: Líneas de la factura.
          items:
            type: object
            properties:
              base_imponible:
                type: string
                description: Base imponible de la línea.
                example: 200
              tipo_impositivo:
                type: string
                description: >
                  Tipo impositivo de la línea. Si `impuesto` = 01 (IVA) entonces
                  los valores permitidos son: 0, 2, 4, 5, 7.5, 10, 21.
                example: '21'
              cuota_repercutida:
                type: string
                description: Cuota repercutida de la linea.
                example: '42'
              impuesto:
                type: string
                default: '01'
                enum:
                  - '01'
                  - '02'
                  - '03'
                  - '05'
                description: |
                  Tipo de impuesto.
                    - 01: Impuesto sobre el Valor Añadido (IVA)
                    - 02: Impuesto sobre la Producción, los Servicios y la Importación (IPSI) de Ceuta y Melilla
                    - 03: Impuesto General Indirecto Canario (IGIC)
                    - 05: Otros
                example: '01'
              calificacion_operacion:
                type: string
                enum:
                  - S1
                  - S2
                  - N1
                  - N2
                description: >
                  Calificación de la operación. Solo se puede informar si el
                  campo `operacion_exenta` no se cumplimenta. En ese caso, por
                  defecto se usará el valor `S1`. 
                    - S1: Operación sujeta y no exenta - sin inversión del sujeto pasivo.
                    - S2: Operación sujeta y no exenta - con inversión del sujeto pasivo.
                    - N1: Operación no sujeta (art. 7, 14, otros).
                    - N2: Operación no sujeta por reglas de localización.
              clave_regimen:
                type: string
                enum:
                  - '01'
                  - '02'
                  - '03'
                  - '04'
                  - '05'
                  - '06'
                  - '07'
                  - '08'
                  - '09'
                  - '10'
                  - '11'
                  - '14'
                  - '15'
                  - '17'
                  - '18'
                  - '19'
                  - '20'
                description: Clave que identifica el tipo de régimen del IVA/IGIC.
              operacion_exenta:
                type: string
                enum:
                  - E1
                  - E2
                  - E3
                  - E4
                  - E5
                  - E6
                description: Tipo de operación exenta.
              base_imponible_a_coste:
                type: string
                description: Base imponible a coste de la linea.
              tipo_recargo_equivalencia:
                type: string
                description: Tipo de recargo de equivalencia de la linea.
              cuota_recargo_equivalencia:
                type: string
                description: Cuota recargo de equivalencia de la linea.
        importe_total:
          description: Importe total de la factura.
          type: string
          example: '242.00'
        cuota_total:
          description: Cuota total de la factura.
          type: string
          example: '42.00'
        encadenamiento:
          type: object
          description: >-
            Datos sobre registro de facturación anterior con el que el registro
            ha sido encadenado.
          properties:
            nif_emisor:
              type: string
              example: B75777847
              description: NIF del emisor del registro de facturación anterior.
            num_serie:
              type: string
              example: A34543
              description: Serie y número del registro de facturación anterior.
            fecha_expedicion:
              type: string
              example: 01-12-2024
              description: Fecha de expedición del registro de facturación anterior.
            huella:
              type: string
              example: 058EE12D8FB181EE51901736CA561F8E084857D976560C421D9F6668251645F3
              description: Huella del registro de facturación anterior.
        huella:
          description: Huella de la factura.
          type: string
          example: F7811516E953893596430BFBC48DD2F790E55DA098E410A430C79EF945E1EAF5
        ultima_modificacion:
          description: Fecha y hora de la última modificación.
          type: string
          example: '2024-12-31T11:30:57+01:00'
        estado:
          description: |
            Estado de la factura en el sistema de la AEAT.
            <ol style="list-style: disc;">
              <li>Correcta: factura registrada correctamente.</li>
              <li>AceptadaConErrores: factura aceptada por la AEAT que contiene errores. Debe ser subsanada emitiendo una nueva factura rectificativa o usando el endpoint POST <code>/modify</code> dependiendo del caso.</li>
              <li>Anulada: factura anulada correctamente.</li>
            </ol>

          type: string
          example: Correcta
...